@@ -1,10 +1,8 @@
 module github.com/kemsta/go-easyrsa/easyrsa-cli
 
-go 1.12
+go 1.14
 
 require (
-	github.com/inconshreveable/mousetrap v1.0.0 // indirect
-	github.com/kemsta/go-easyrsa v0.2.4
-	github.com/spf13/cobra v0.0.3
-	github.com/spf13/pflag v1.0.3 // indirect
+	github.com/kemsta/go-easyrsa v0.2.5
+	github.com/spf13/cobra v0.0.7
 )
