@@ -1,16 +1,9 @@
 module github.com/kemsta/go-easyrsa
 
-go 1.12
+go 1.14
 
 require (
 	github.com/gofrs/flock v0.7.1
-	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
-	github.com/kr/pretty v0.1.0 // indirect
-	github.com/pkg/errors v0.8.1
-	github.com/prometheus/common v0.2.0
-	github.com/sirupsen/logrus v1.4.0 // indirect
-	github.com/stretchr/testify v1.3.0
-	golang.org/x/crypto v0.0.0-20190320223903-b7391e95e576 // indirect
-	golang.org/x/sys v0.0.0-20190321052220-f7bb7a8bee54 // indirect
-	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
+	github.com/prometheus/common v0.9.1
+	github.com/stretchr/testify v1.5.1
 )
